export rgToHex fn
